Chat Bot using Python
A chatbot is a computer program or AI that can interact with humans. It responds to human input and can answer questions, carry on conversations, provide information, and perform tasks. The goal of a chatbot is to provide a conversational and engaging experience for users.

Using Python, you can create a chatbot that will be able to accomplish these tasks. First, you need to create a function that can take in a user input and then output a response. For example, you can create a function that will check whether the user's input matches certain phrases and then output a response.

Once you’ve created a function to interpret the user input, you can create a loop in which the chatbot will keep running. Through this loop, it can keep track of the user’s conversation and respond accordingly. You can also include additional features such as natural language processing to make it understand more complex queries. 

If you’re looking to build on the basic features of a chatbot, you can also use libraries such as NLP to create a more advanced bot. NLP is a powerful tool for understanding human language and can be used for creating more sophisticated conversations.